requirements.txt
matplotlib==2.2.3
numba==0.39.0
numpy-quaternion==2018.7.5.21.55.13
opencv-python==3.4.3.18
scikit-image==0.14.0
scikit-learn==0.19.2
scikit-video==1.1.10
svgwrite==1.1.12
tables==3.4.4
ujson==1.35
jsonschema==2.6.0
scipy==1.1.0
numpy==1.14.5
